Transition (#35)
* Transition

fades in and scrolls to the top as needed

* fixed issues

should be good now

* making the transition work

have it where the new page fades in properly.

* almost fixed

I got this to the point where there is a cross dissolve, but the only issue is that the old page scrolls to the top and fades out while the second one fades in. wil figure it out soon

* feat: Resolve conflicts

* more fixed

the animation looks smoother but i still cant get rid of the jump

* more updated (not final)

* feat: Add dependencies

* refactor: Unused import

---------

Co-authored-by: Hanyi Liu <35943468+hanyiliu@users.noreply.github.com>
Co-authored-by: Hanyi Liu <hanyiliu0@gmail.com>